# docs/swagger.yaml
openapi: 3.0.3
info:
  title: FanFare API
  description: |
    Pop Culture Prediction & Sentiment Platform Backend API.
    
    FanFare enables fans to predict outcomes, track community sentiment,
    contribute crowdsourced context cards, and build reputation through
    accurate insights — not follower counts.
    
    ## 4 Key Product Pillars
    1. **Predictions Drive Engagement** — Users predict box office, awards, ratings, story events
    2. **Sentiment Dashboards Show the Fandom Pulse** — Track hype, satisfaction, controversies over time
    3. **Context Cards Reduce Misinformation** — Structured crowd-sourced notes clarify truth vs rumor
    4. **Reputation from Insight, Not Followers** — Points for accurate predictions & helpful contributions
  version: 1.0.0
  contact:
    name: FanFare Team
servers:
  - url: http://localhost:8080
    description: Local development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/v1/auth/login
  schemas:
...
    PredictionVoteRequest:
      type: object
      required: [selected_option, confidence]
      properties:
        selected_option:
          type: string
          example: "Yes (Above $357M)"
        confidence:
          type: string
          enum: [low, medium, high]
          example: "high"
    PredictionResolveRequest:
      type: object
      required: [correct_option, source]
      properties:
        correct_option:
          type: string
        source:
          type: string
          example: "Box Office Mojo official numbers"
    PredictionParticipation:
      type: object
      properties:
        id:
          type: string
        prediction_id:
          type: string
        user_id:
          type: string
        selected_option:
          type: string
        confidence:
          type: string
        points_awarded:
          type: integer
        is_correct:
          type: boolean
...
    SentimentVoteRequest:
      type: object
      required: [selected_option, region, age_band]
      properties:
        selected_option:
          type: string
          example: "Massively Hyped"
        region:
          type: string
          example: "North America"
        age_band:
          type: string
          example: "Gen Z (18-25)"
...
paths:
...
  /api/v1/predictions/{id}/vote:
    post:
      summary: Place a prediction vote
      tags: [Predictions]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PredictionVoteRequest'
      responses:
        '201':
          description: Vote placed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PredictionParticipation'
        '400':
          description: Invalid request or prediction closed
        '409':
          description: User already voted
  /api/v1/predictions/{id}/resolve:
    post:
      summary: Resolve a prediction (admin)
      tags: [Predictions]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PredictionResolveRequest'
      responses:
        '200':
          description: Prediction resolved and scores calculated
        '400':
          description: Invalid request or already resolved
  /api/v1/sentiment/dashboards/{target_type}/{target_id}:
    get:
      summary: Get sentiment dashboard with aggregated metrics
      tags: [Sentiment]
      security:
        - BearerAuth: []
      parameters:
        - name: target_type
          in: path
          required: true
          schema:
            type: string
            enum: [issue, title]
        - name: target_id
          in: path
          required: true
          schema:
            type: string
        - name: cohort
          in: query
          schema:
            type: string
            enum: [region, age_band]
      responses:
        '200':
          description: Aggregated sentiment distributions
        '401':
          description: Unauthorized — valid JWT required
  /api/v1/sentiment/polls/{id}/vote:
    post:
      summary: Submit a sentiment poll vote
      tags: [Sentiment]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SentimentVoteRequest'
      responses:
        '201':
          description: Vote submitted
        '409':
          description: Already voted
...